requirements.txt
# requirements.txt
flask==2.0.1
spotipy==2.19.0
python-dotenv==0.19.0